backend/requirements.txt

asgiref==3.4.1
astroid==2.5
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.9
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mysql-connector-python==8.0.26
Pillow==9.0.0
protobuf==3.19.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.0.1
pylint==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2020.5
requests==2.8.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==6.0.0
wrapt==1.12.1